pyproject.toml
[project]
name = "stm-data-processing"
version = "0.3.0"
description = "STM / DFT / Wannier90 data processing toolkit"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "hunfen", email = "hunfen@anywhere.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=2.3.0",
    "scipy>=1.16.0",
    "matplotlib>=3.10.0",
    "h5py>=3.15.0",
    "pandas>=2.3.0",
    "tqdm>=4.67.0",
    "mpi4py>=4.1.1",
    "ipykernel>=7.2.0",
    "mpmath>=1.4.0",
    "psutil>=7.2.2",
    "pyfftw>=0.15.1",
]

[project.optional-dependencies]
gpu = ["cupy-cuda13x>=14.0.0",]
dev = [
    "ruff>=0.15.1",
    "ipykernel>=7.2.0",
]

[project.urls]
Repository = "https://github.com/Hunfen/STM_DataProcessing.git"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["stm_data_processing*"]
exclude = ["tests*", "docs*"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle error
    "W",   # pycodestyle warning  
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "YTT", # flake8-2020
    "Q",   # flake8-quotes
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    "RUF", # Ruff
]
ignore = [
    "E501"
    ]
per-file-ignores = {"__init__.py" = ["F401"]}

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
